Refactoring of message encoding/decoding. All taken care of by code-generator now. Makes the code a lot simplier to follow.

+++ src/MiNET/MiNET.Network/RakNetIdentifiers.cs
@@ -38,7 +38,7 @@
 	//typedef double MicrosecondsPerByte;
 
 
-	public enum DefaultMessageIdTypes: int
+	public enum DefaultMessageIdTypes: byte
 	{
 		//
 		// RESERVED TYPES - DO NOT CHANGE THESE
